+++ 2024/src/app/home/_components/Sponsor.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import SponsorSwiper from "./SponsorSwiper";
+
+const Sponsor = () => {
+  return (
+    <section className="pt-16 pb-20 mb-32">
+      <h4 className="block-title text-center mb-12">贊助夥伴</h4>
+      <SponsorSwiper />
+    </section>
+  );
+};
+
+export default Sponsor;
