+++ 2024/src/app/home/page.jsx
@@ -1,0 +1,18 @@
+import React, { Fragment } from "react";
+import Hero from "./_components/Hero";
+import About from "./_components/About";
+import Venue from "./_components/Venue";
+import Transportation from "./_components/Transportation";
+
+const Home = () => {
+  return (
+    <Fragment>
+      <Hero />
+      <About />
+      <Venue />
+      <Transportation />
+    </Fragment>
+  );
+};
+
+export default Home;
